Move definitions from bridge to worker
