+++ temporalio/worker/_tuning.py
@@ -3,22 +3,11 @@ import logging
 from abc import ABC, abstractmethod
 from dataclasses import dataclass
 from datetime import timedelta
-from typing import Any, Callable, Literal, Optional, Union
+from typing import Any, Callable, Literal, Optional, Protocol, Union, runtime_checkable
 
 from typing_extensions import TypeAlias
 
 import temporalio.bridge.worker
-from temporalio.bridge.worker import (
-    ActivitySlotInfo,
-    CustomSlotSupplier,
-    LocalActivitySlotInfo,
-    SlotInfo,
-    SlotMarkUsedContext,
-    SlotPermit,
-    SlotReleaseContext,
-    SlotReserveContext,
-    WorkflowSlotInfo,
-)
 
 _DEFAULT_RESOURCE_ACTIVITY_MAX = 500
 
@@ -83,6 +72,134 @@ class ResourceBasedSlotSupplier:
     tuner_config: ResourceBasedTunerConfig
     """Options for the tuner that will be used to adjust the number of slots. When used with a
     :py:class:`CompositeTuner`, all resource-based slot suppliers must use the same tuner options."""
+
+
+class SlotPermit:
+    """A permit to use a slot for a workflow/activity/local activity task.
+
+    You can inherit from this class to add your own data to the permit.
+    """
+
+    pass
+
+
+# WARNING: This must match Rust worker::SlotReserveCtx
+class SlotReserveContext(Protocol):
+    """Context for reserving a slot from a :py:class:`CustomSlotSupplier`."""
+
+    slot_type: Literal["workflow", "activity", "local-activity"]
+    """The type of slot trying to be reserved. Always one of "workflow", "activity", or "local-activity"."""
+    task_queue: str
+    """The name of the task queue for which this reservation request is associated."""
+    worker_identity: str
+    """The identity of the worker that is requesting the reservation."""
+    worker_build_id: str
+    """The build id of the worker that is requesting the reservation."""
+    is_sticky: bool
+    """True iff this is a reservation for a sticky poll for a workflow task."""
+
+
+@runtime_checkable
+class WorkflowSlotInfo(Protocol):
+    """Info about a workflow task slot usage."""
+
+    workflow_type: str
+    is_sticky: bool
+
+
+@runtime_checkable
+class ActivitySlotInfo(Protocol):
+    """Info about an activity task slot usage."""
+
+    activity_type: str
+
+
+@runtime_checkable
+class LocalActivitySlotInfo(Protocol):
+    """Info about a local activity task slot usage."""
+
+    activity_type: str
+
+
+SlotInfo: TypeAlias = Union[WorkflowSlotInfo, ActivitySlotInfo, LocalActivitySlotInfo]
+
+
+# WARNING: This must match Rust worker::SlotMarkUsedCtx
+class SlotMarkUsedContext(Protocol):
+    """Context for marking a slot used from a :py:class:`CustomSlotSupplier`."""
+
+    slot_info: SlotInfo
+    """Info about the task that will be using the slot."""
+    permit: SlotPermit
+    """The permit that was issued when the slot was reserved."""
+
+
+@dataclass(frozen=True)
+class SlotReleaseContext:
+    """Context for releasing a slot from a :py:class:`CustomSlotSupplier`."""
+
+    slot_info: SlotInfo
+    """Info about the task that will be using the slot."""
+    permit: SlotPermit
+    """The permit that was issued when the slot was reserved."""
+
+
+class CustomSlotSupplier(ABC):
+    """This class can be implemented to provide custom slot supplier behavior."""
+
+    @abstractmethod
+    async def reserve_slot(self, ctx: SlotReserveContext) -> SlotPermit:
+        """This function is called before polling for new tasks. Your implementation must block until a
+        slot is available then return a permit to use that slot.
+
+        The only acceptable exception to throw is :py:class:`asyncio.CancelledError`, as invocations of this method may
+        be cancelled. Any other exceptions thrown will be logged and ignored.
+
+        Args:
+            ctx: The context for slot reservation.
+
+        Returns:
+            A permit to use the slot which may be populated with your own data.
+        """
+        ...
+
+    @abstractmethod
+    def try_reserve_slot(self, ctx: SlotReserveContext) -> Optional[SlotPermit]:
+        """This function is called when trying to reserve slots for "eager" workflow and activity tasks.
+        Eager tasks are those which are returned as a result of completing a workflow task, rather than
+        from polling. Your implementation must not block, and if a slot is available, return a permit
+        to use that slot.
+
+        Args:
+            ctx: The context for slot reservation.
+
+        Returns:
+            Maybe a permit to use the slot which may be populated with your own data.
+        """
+        ...
+
+    @abstractmethod
+    def mark_slot_used(self, ctx: SlotMarkUsedContext) -> None:
+        """This function is called once a slot is actually being used to process some task, which may be
+        some time after the slot was reserved originally. For example, if there is no work for a
+        worker, a number of slots equal to the number of active pollers may already be reserved, but
+        none of them are being used yet. This call should be non-blocking.
+
+        Args:
+            ctx: The context for marking a slot as used.
+        """
+        ...
+
+    @abstractmethod
+    def release_slot(self, ctx: SlotReleaseContext) -> None:
+        """This function is called once a permit is no longer needed. This could be because the task has
+        finished, whether successfully or not, or because the slot was no longer needed (ex: the number
+        of active pollers decreased). This call should be non-blocking.
+
+        Args:
+            ctx: The context for releasing a slot.
+        """
+        ...
 
 
 SlotSupplier: TypeAlias = Union[
